+++ NEWS.md
@@ -1,3 +1,7 @@
+# verison 0.7.1
+
+- Fix include_xy and include_cell arguments for terra inputs
+
 # version 0.7.0
 
 - Support terra SpatRaster inputs